nasa-ai-model/requirements.txt
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
gradio>=3.40.0
numpy>=1.24.0
pyyaml>=6.0
